requirements.txt
anthropic==0.49.0
openai==1.63.0
protobuf==4.25.3
torch==2.1.2
transformers==4.43.1
google-genai== 1.2.0
numpy==1.26.4 
accelerate==1.4.0
tiktoken==0.9.0
peft==0.12.0
pandas==2.2.0